@@ -1,15 +1,22 @@
 {
-  "name": "new",
+  "name": "react-redux-hooks-jwt-auth",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^5.0.0",
-    "gotrue-js": "^0.9.29",
-    "react": "^16.8.6",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^16.8.6",
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
+    "axios": "^0.21.1",
+    "bootstrap": "^4.5.2",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-redux": "^7.2.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3"
+    "react-scripts": "3.4.3",
+    "react-validation": "^3.0.7",
+    "redux": "^4.0.5",
+    "redux-thunk": "^2.3.0",
+    "validator": "^13.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -31,5 +38,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "redux-devtools-extension": "^2.13.8"
   }
 }
